requirements.txt
pyserial>=3.5
SpeechRecognition>=3.8
PyAudio>=0.2.11